+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## BREAKING CHANGES
 
-* `fq_subsample`: This component has been removed in favour of `fq/fq_subsample` (PR #xxx).
+* `fq_subsample`: This component has been removed in favour of `fq/fq_subsample` (PR #182).
 
 ## NEW FUNCTIONALITY
 
 * `fq`:
-  - `fq/fq_filter`: Filter FASTQ files based on record names or sequence patterns (PR #xxx).
+  - `fq/fq_filter`: Filter FASTQ files based on record names or sequence patterns (PR #182).
 
 # biobox 0.3.2